meh : minor type change... unless we absolutely did not want to support more than 2147483647 characters long formulas

+++ slnCpp_Challenges/MathParser2/FormulaParser.cpp
@@ -99,7 +99,7 @@ int FormulaParser::splitFormula(const string pMain, string* pSubFormula1, string
 {
 	for (vector<P_Expression>::iterator it = supportedExpressions->begin(); it != supportedExpressions->end(); ++it)
 	{
-		for (int i = pMain.length(); i >= 0; --i)
+		for (size_t i = pMain.length(); i >= 0; --i)
 		{
 			if (pMain[i] == ')')
 			{
